Remove import cycle for query history domain mappers
